Move calculation methods into sub-submodule
